got game lobby to work and propagate
