+++ src/main.tsx
@@ -5,6 +5,7 @@ import App from "./App.tsx";
 import { createBrowserRouter, RouterProvider } from "react-router";
 import { TicTacToeApiClient } from "./api.ts";
 import Lobby from "./Lobby.tsx";
+import GameView from "./GameView.tsx";
 
 const api = new TicTacToeApiClient();
 
@@ -19,6 +20,20 @@ const router = createBrowserRouter([
         loader: async () => {
           const games = await api.getGames();
           return { games };
+        },
+      },
+      {
+        path: "/game/:gameId",
+        Component: GameView,
+        loader: async ({ params }) => {
+          if (params.gameId) {
+            
+            const game = await api.getGame(params.gameId);
+            return { game };
+          } else {
+            console.error()
+            return null
+          }
         },
       },
     ],
